Small refactor for music constants
