--- src/components/music/helpers/music_constants.js
@@ -1,185 +1,0 @@
-export const C_3 = 0;
-export const Cs3 = 1;
-export const D_3 = 2;
-export const Ds3 = 3;
-export const E_3 = 4;
-export const F_3 = 5;
-export const Fs3 = 6;
-export const G_3 = 7;
-export const Gs3 = 8;
-export const A_3 = 9;
-export const As3 = 10;
-export const B_3 = 11;
-export const C_4 = 12;
-export const Cs4 = 13;
-export const D_4 = 14;
-export const Ds4 = 15;
-export const E_4 = 16;
-export const F_4 = 17;
-export const Fs4 = 18;
-export const G_4 = 19;
-export const Gs4 = 20;
-export const A_4 = 21;
-export const As4 = 22;
-export const B_4 = 23;
-export const C_5 = 24;
-export const Cs5 = 25;
-export const D_5 = 26;
-export const Ds5 = 27;
-export const E_5 = 28;
-export const F_5 = 29;
-export const Fs5 = 30;
-export const G_5 = 31;
-export const Gs5 = 32;
-export const A_5 = 33;
-export const As5 = 34;
-export const B_5 = 35;
-export const C_6 = 36;
-export const Cs6 = 37;
-export const D_6 = 38;
-export const Ds6 = 39;
-export const E_6 = 40;
-export const F_6 = 41;
-export const Fs6 = 42;
-export const G_6 = 43;
-export const Gs6 = 44;
-export const A_6 = 45;
-export const As6 = 46;
-export const B_6 = 47;
-export const C_7 = 48;
-export const Cs7 = 49;
-export const D_7 = 50;
-export const Ds7 = 51;
-export const E_7 = 52;
-export const F_7 = 53;
-export const Fs7 = 54;
-export const G_7 = 55;
-export const Gs7 = 56;
-export const A_7 = 57;
-export const As7 = 58;
-export const B_7 = 59;
-export const C_8 = 60;
-export const Cs8 = 61;
-export const D_8 = 62;
-export const Ds8 = 63;
-export const E_8 = 64;
-export const F_8 = 65;
-export const Fs8 = 66;
-export const G_8 = 67;
-export const Gs8 = 68;
-export const A_8 = 69;
-export const As8 = 70;
-export const B_8 = 71;
-
-export const noteNames = [
-  "C-3",
-  "C#3",
-  "D-3",
-  "D#3",
-  "E-3",
-  "F-3",
-  "F#3",
-  "G-3",
-  "G#3",
-  "A-3",
-  "A#3",
-  "B-3",
-  "C-4",
-  "C#4",
-  "D-4",
-  "D#4",
-  "E-4",
-  "F-4",
-  "F#4",
-  "G-4",
-  "G#4",
-  "A-4",
-  "A#4",
-  "B-4",
-  "C-5",
-  "C#5",
-  "D-5",
-  "D#5",
-  "E-5",
-  "F-5",
-  "F#5",
-  "G-5",
-  "G#5",
-  "A-5",
-  "A#5",
-  "B-5",
-  "C-6",
-  "C#6",
-  "D-6",
-  "D#6",
-  "E-6",
-  "F-6",
-  "F#6",
-  "G-6",
-  "G#6",
-  "A-6",
-  "A#6",
-  "B-6",
-  "C-7",
-  "C#7",
-  "D-7",
-  "D#7",
-  "E-7",
-  "F-7",
-  "F#7",
-  "G-7",
-  "G#7",
-  "A-7",
-  "A#7",
-  "B-7",
-  "C-8",
-  "C#8",
-  "D-8",
-  "D#8",
-  "E-8",
-  "F-8",
-  "F#8",
-  "G-8",
-  "G#8",
-  "A-8",
-  "A#8",
-  "B-8",
-];
-
-export const note2freq = [
-  /*C_3*/ 44, /*Cs3*/ 156, /*D_3*/ 262, /*Ds3*/ 363, /*E_3*/ 457, /*F_3*/ 547,
-  /*Fs3*/ 631, /*G_3*/ 710, /*Gs3*/ 786, /*A_3*/ 854, /*As3*/ 923, /*B_3*/ 986,
-  /*C_4*/ 1046, /*Cs4*/ 1102, /*D_4*/ 1155, /*Ds4*/ 1205, /*E_4*/ 1253,
-  /*F_4*/ 1297, /*Fs4*/ 1339, /*G_4*/ 1379, /*Gs4*/ 1417, /*A_4*/ 1452,
-  /*As4*/ 1486, /*B_4*/ 1517, /*C_5*/ 1546, /*Cs5*/ 1575, /*D_5*/ 1602,
-  /*Ds5*/ 1627, /*E_5*/ 1650, /*F_5*/ 1673, /*Fs5*/ 1694, /*G_5*/ 1714,
-  /*Gs5*/ 1732, /*A_5*/ 1750, /*As5*/ 1767, /*B_5*/ 1783, /*C_6*/ 1798,
-  /*Cs6*/ 1812, /*D_6*/ 1825, /*Ds6*/ 1837, /*E_6*/ 1849, /*F_6*/ 1860,
-  /*Fs6*/ 1871, /*G_6*/ 1881, /*Gs6*/ 1890, /*A_6*/ 1899, /*As6*/ 1907,
-  /*B_6*/ 1915, /*C_7*/ 1923, /*Cs7*/ 1930, /*D_7*/ 1936, /*Ds7*/ 1943,
-  /*E_7*/ 1949, /*F_7*/ 1954, /*Fs7*/ 1959, /*G_7*/ 1964, /*Gs7*/ 1969,
-  /*A_7*/ 1974, /*As7*/ 1978, /*B_7*/ 1982, /*C_8*/ 1985, /*Cs8*/ 1988,
-  /*D_8*/ 1992, /*Ds8*/ 1995, /*E_8*/ 1998, /*F_8*/ 2001, /*Fs8*/ 2004,
-  /*G_8*/ 2006, /*Gs8*/ 2009, /*A_8*/ 2011, /*As8*/ 2013, /*B_8*/ 2015,
-];
-
-// Sound controller registers
-export const NR10 = 0xff10;
-export const NR11 = 0xff11;
-export const NR12 = 0xff12;
-export const NR13 = 0xff13;
-export const NR14 = 0xff14;
-export const NR21 = 0xff16;
-export const NR22 = 0xff17;
-export const NR23 = 0xff18;
-export const NR24 = 0xff19;
-export const NR30 = 0xff1a;
-export const NR31 = 0xff1b;
-export const NR32 = 0xff1c;
-export const NR33 = 0xff1d;
-export const NR34 = 0xff1e;
-export const NR41 = 0xff20;
-export const NR42 = 0xff21;
-export const NR43 = 0xff22;
-export const NR44 = 0xff23;
-export const AUD3_WAVE_RAM = 0xff30;
